Added Dissolve VFX
Added the Shaders plugin and Dissolve Effect into Unity. I also added a card texture so the Dissolve could be used on something.

For some reason the card turns dark red on the Game Preview window. If it's messing with your cards, simply go to MoveCard > CardBackground and remove the material (and turn the sprite back to Background if you don't like the card reference).